PS-56 - Porting prototype into production React SPA with fixture-backed hooks
