+++ resources/js/hooks/use-exercises.ts
@@ -1,4 +1,4 @@
-import { useApp } from '@/lib/store'
+import { useApp } from '@/lib/use-app'
 import type { Exercise } from '@/api/types'
 
 interface UseExercisesResult {
