+++ resources/js/hooks/use-profile.ts
@@ -1,4 +1,4 @@
-import { useApp } from '@/lib/store'
+import { useApp } from '@/lib/use-app'
 import type { User } from '@/api/types'
 
 interface UseProfileResult {
