+++ resources/js/lib/store.tsx
@@ -1,4 +1,4 @@
-import { createContext, useCallback, useContext, useEffect, useState, type ReactNode } from 'react'
+import { useCallback, useEffect, useState, type ReactNode } from 'react'
 import type {
   User,
   EquipmentType,
@@ -14,44 +14,11 @@ import {
   fixtureWorkouts,
   fixtureTemplates,
 } from '@/api/fixtures'
+import { AppContext, type AppContextValue } from './app-context'
 
 interface Toast {
   id: string
   message: string
-}
-
-interface AppContextValue {
-  user: User
-  authed: boolean
-  exercises: Exercise[]
-  equipment: EquipmentType[]
-  workouts: Workout[]
-  templates: WorkoutTemplate[]
-  toasts: Toast[]
-  uid: (prefix: string) => string
-  toast: (message: string) => void
-  updateUser: (patch: Partial<User>) => void
-  login: () => void
-  logout: () => void
-  addExercise: (ex: Partial<Exercise> & { name: string; type: Exercise['type'] }) => Exercise
-  updateExercise: (id: string, patch: Partial<Exercise>) => void
-  deleteExercise: (id: string) => void
-  addEquipment: (name: string) => EquipmentType
-  deleteEquipment: (id: string) => void
-  addWorkout: (w: Partial<Workout> & { name: string; date: string }) => Workout
-  updateWorkout: (id: string, patch: Partial<Workout>) => void
-  deleteWorkout: (id: string) => void
-  createFromTemplate: (templateId: string, date: string, name?: string) => Workout | null
-  updateTemplate: (id: string, patch: Partial<WorkoutTemplate>) => void
-  deleteTemplate: (id: string) => void
-}
-
-const AppContext = createContext<AppContextValue | null>(null)
-
-export function useApp(): AppContextValue {
-  const ctx = useContext(AppContext)
-  if (!ctx) throw new Error('useApp must be used within AppProvider')
-  return ctx
 }
 
 function clone<T>(x: T): T {
