+++ resources/js/lib/use-app.ts
@@ -1,0 +1,8 @@
+import { useContext } from 'react'
+import { AppContext, type AppContextValue } from './app-context'
+
+export function useApp(): AppContextValue {
+  const ctx = useContext(AppContext)
+  if (!ctx) throw new Error('useApp must be used within AppProvider')
+  return ctx
+}
